Move <HorizontalSeparator> into layoutComponents
