+++ ui/src/components/stop-registry/stops/stop-details/layout/index.ts
@@ -1,4 +1,3 @@
 export * from './DetailRow';
-export * from './HorizontalSeparator';
 export * from './LabeledDetail';
 export * from './SlimSimpleButton';
