+++ ui/src/layoutComponents/index.ts
@@ -1,4 +1,5 @@
 export * from './Column';
 export * from './Container';
 export * from './Row';
+export * from './HorizontalSeparator';
 export * from './Visible';
